# docs/swagger.yaml
basePath: /
definitions:
  ttlock.AccessTokenRequestParams:
    properties:
      clientId:
        type: string
      clientSecret:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - clientId
    - clientSecret
    - password
    - username
    type: object
...
  ttlock.PasscodeCustomRequestParams:
    properties:
      accessToken:
        type: string
      addType:
        enum:
        - 1
        - 2
        type: integer
      clientId:
        type: string
      date:
        type: integer
      endDate:
        type: integer
      keyboardPwd:
        type: integer
      keyboardPwdName:
        type: string
      keyboardPwdType:
        enum:
        - 2
        - 3
        type: integer
      lockId:
        type: integer
      startDate:
        type: integer
    required:
    - accessToken
    - addType
    - clientId
    - date
    - endDate
    - keyboardPwd
    - keyboardPwdType
    - lockId
    - startDate
    type: object
  ttlock.PasscodeDeleteRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      deleteType:
        enum:
        - 1
        - 2
        type: integer
      keyboardPwdId:
        type: integer
      lockId:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - deleteType
    - keyboardPwdId
    - lockId
    type: object
...
  ttlock.PasscodeRandomRequestParams:
    properties:
      accessToken:
        type: string
      clientId:
        type: string
      date:
        type: integer
      endDate:
        type: integer
      keyboardPwdName:
        type: string
      keyboardPwdType:
        type: integer
      lockId:
        type: integer
      startDate:
        type: integer
    required:
    - accessToken
    - clientId
    - date
    - endDate
    - keyboardPwdType
    - lockId
    - startDate
    type: object
  ttlock.PasscodeUpdateRequestParams:
    properties:
      accessToken:
        type: string
      changeType:
        enum:
        - 1
        - 2
        type: integer
      clientId:
        type: string
      date:
        type: integer
      endDate:
        type: integer
      keyboardPwdId:
        type: integer
      keyboardPwdName:
        type: string
      lockId:
        type: integer
      newKeyboardPwd:
        type: string
      startDate:
        type: integer
    required:
    - accessToken
    - changeType
    - clientId
    - date
    - keyboardPwdId
    - lockId
    type: object
  ttlock.RefreshAccessTokenRequestParams:
    properties:
      clientId:
        type: string
      clientSecret:
        type: string
      grant_type:
        type: string
      refresh_token:
        type: string
    required:
    - clientId
    - clientSecret
    - grant_type
    - refresh_token
    type: object
  ttlock.UserDeleteRequestParams:
    properties:
      clientId:
        type: string
      clientSecret:
        type: string
      date:
        type: integer
      username:
        type: string
    required:
    - clientId
    - clientSecret
    - date
    - username
    type: object
  ttlock.UserListRequestParams:
    properties:
      clientId:
        type: string
      clientSecret:
        type: string
      date:
        type: integer
      endDate:
        type: integer
      pageNo:
        minimum: 1
        type: integer
      pageSize:
        maximum: 100
        minimum: 1
        type: integer
      startDate:
        type: integer
      username:
        type: string
    required:
    - clientId
    - clientSecret
    - date
    - pageNo
    - pageSize
    type: object
  ttlock.UserPasswordResetRequestParams:
    properties:
      clientId:
        type: string
      clientSecret:
        type: string
      date:
        type: integer
      password:
        type: string
      username:
        type: string
    required:
    - clientId
    - clientSecret
    - date
    - password
    - username
    type: object
  ttlock.UserRegisterRequestParams:
    properties:
      clientId:
        type: string
      clientSecret:
        type: string
      date:
        type: integer
      password:
        type: string
      username:
        type: string
    required:
    - clientId
    - clientSecret
    - date
    - password
    - username
    type: object
info:
  contact: {}
  title: TTLock Api
  version: "1.0"
paths:
  /api/auth/access-token:
    post:
      consumes:
      - application/json
      description: Generate TTLock access token
      parameters:
      - description: Access token request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.AccessTokenRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Access token generated successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get access token
      tags:
      - Authentication
  /api/auth/refresh-access-token:
    post:
      consumes:
      - application/json
      description: Refresh TTLock access token using refresh token
      parameters:
      - description: Refresh access token request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.RefreshAccessTokenRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Access token refreshed successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Refresh access token
      tags:
      - Authentication
...
  /api/passcode/delete:
    post:
      consumes:
      - application/json
      description: Delete a TTLock passcode
      parameters:
      - description: Passcode delete request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.PasscodeDeleteRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Passcode deleted successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Delete passcode
      tags:
      - Passcodes
  /api/passcode/generate-custom:
    post:
      consumes:
      - application/json
      description: Generate a custom TTLock passcode
      parameters:
      - description: Custom passcode request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.PasscodeCustomRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Custom passcode generated successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Generate custom passcode
      tags:
      - Passcodes
  /api/passcode/generate-random:
    post:
      consumes:
      - application/json
      description: Generate a random TTLock passcode
      parameters:
      - description: Random passcode request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.PasscodeRandomRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Random passcode generated successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Generate random passcode
      tags:
      - Passcodes
  /api/passcode/types:
    post:
      consumes:
      - application/json
      description: Get all TTLock passcode types
      produces:
      - application/json
      responses:
        "200":
          description: Passcode types fetched successfully
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get passcode types
      tags:
      - Passcodes
  /api/passcode/update:
    post:
      consumes:
      - application/json
      description: Update a TTLock passcode
      parameters:
      - description: Passcode update request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.PasscodeUpdateRequestParams'
      produces:
      - application/json
      responses:
        "200":
          description: Passcode updated successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Update passcode
      tags:
      - Passcodes
  /api/users/delete:
    post:
      consumes:
      - application/json
      description: Delete a TTLock user
      parameters:
      - description: User delete request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.UserDeleteRequestParams'
      produces:
      - application/json
      responses:
        "202":
          description: User deleted successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Delete user
      tags:
      - Users
  /api/users/list:
    post:
      consumes:
      - application/json
      description: Get a list of TTLock users
      parameters:
      - description: Users list request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.UserListRequestParams'
      produces:
      - application/json
      responses:
        "202":
          description: Users list fetched successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Get users list
      tags:
      - Users
  /api/users/password-reset:
    post:
      consumes:
      - application/json
      description: Reset TTLock user password
      parameters:
      - description: User password reset request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.UserPasswordResetRequestParams'
      produces:
      - application/json
      responses:
        "202":
          description: User password reset successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Reset user password
      tags:
      - Users
  /api/users/register:
    post:
      consumes:
      - application/json
      description: Register a new TTLock user
      parameters:
      - description: User register request params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ttlock.UserRegisterRequestParams'
      produces:
      - application/json
      responses:
        "202":
          description: User registered successfully
          schema: {}
        "400":
          description: Bad request
          schema: {}
        "500":
          description: Internal server error
          schema: {}
      summary: Register user
      tags:
      - Users
swagger: "2.0"